fix(grid): removed old generated sample and changed name of the correct one
