+++ browser/IgBlazorSamples.Client/wwwroot/index.html
@@ -119,7 +119,6 @@
 <script src="sb/grid-data-validator-service-extended.js"></script>
 <script src="sb/grid-editing-events.js"></script>
 <script src="sb/grid-editing-excel-style.js"></script>
-<script src="sb/grid-editing-excel-style-custom.js"></script>
 <script src="sb/grid-editing-lifecycle.js"></script>
 <script src="sb/grid-excel-style-filtering-sample-1.js"></script>
 <script src="sb/grid-excel-style-filtering-sample-2.js"></script>
