--- samples/grids/grid/editing-excel-style-custom/BlazorClientApp.csproj
@@ -1,21 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.BlazorWebAssembly">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <RazorLangVersion>3.0</RazorLangVersion>
-    <AssemblyName>Infragistics.Samples</AssemblyName>
-    <RootNamespace>Infragistics.Samples</RootNamespace>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <NoWarn>1701;1702,IDE0028,BL0005,0219,CS1998</NoWarn>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="IgniteUI.Blazor" Version="24.1.67" />
-    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="8.0.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="8.0.0" />
-    <PackageReference Include="System.Net.Http.Json" Version="8.0.0" />
-  </ItemGroup>
-
-</Project>
