--- samples/grids/grid/editing-excel-style-custom/Program.cs
@@ -1,37 +1,0 @@
-using System;
-using System.Net.Http;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using System.Text;
-using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using IgniteUI.Blazor.Controls;
-using System.Text.Json;
-using System.Text.Json.Serialization;
-
-namespace Infragistics.Samples
-{
-    public class Program
-    {
-        public static async Task Main(string[] args)
-        {
-            var builder = WebAssemblyHostBuilder.CreateDefault(args);
-            builder.RootComponents.Add<App>("app");
-            builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
-            // registering Ignite UI modules
-            builder.Services.AddIgniteUIBlazor(
-                typeof(IgbInputModule),
-                typeof(IgbGridModule)
-            );
-
-            builder.Services.Configure<JsonSerializerOptions>(options =>
-            {
-                options.ReferenceHandler = ReferenceHandler.Preserve;
-                options.MaxDepth = 64; // Increase the maximum depth if needed
-            });
-            await builder.Build().RunAsync();
-        }
-    }
-}
