--- samples/grids/grid/editing-excel-style-custom/ReadMe.md
@@ -1,70 +1,0 @@
-<!-- NOTE: do not change this file because it's auto re-generated from template: -->
-<!-- https://github.com/IgniteUI/igniteui-blazor-examples/tree/vnext/templates/sample/ReadMe.md -->
-
-This folder contains implementation of Blazor application with example of Editing Excel Style Custom feature using [Grid](https://www.infragistics.com/products/ignite-ui-blazor/blazor/components/general-getting-started.html) component.
-
-
-<html lang="en" xmlns="http://www.w3.org/1999/xhtml">
-    <body>
-        <!-- https://static.infragistics.com/xplatform/images/browsers -->
-        <a target="_blank" href="https://www.infragistics.com/products/ignite-ui-blazor/blazor/components/general-getting-started.html" rel="noopener noreferrer">
-            <img height="40px" style="border-radius: 0rem" alt="View Docs" src="https://github.com/IgniteUI/igniteui-blazor-examples/raw/vnext/templates/sample/images/button-docs.png"/>
-        </a>
-        <a target="_blank" href="./App.razor" rel="noopener noreferrer">
-            <img height="40px" style="border-radius: 0rem; max-width: 100%;" alt="View Code" src="https://github.com/IgniteUI/igniteui-blazor-examples/raw/vnext/templates/sample/images/button-code.png"/>
-        </a>
-        <a target="_blank" href="https://infragistics.com/blazor-client/samples/grids/grid/editing-excel-style-custom" rel="noopener noreferrer">
-            <img height="40px" style="border-radius: 0rem; max-width: 100%;" alt="Run Sample" src="https://github.com/IgniteUI/igniteui-blazor-examples/raw/vnext/templates/sample/images/button-run.png"/>
-        </a>
-        <!-- <a target="_blank" href="https://codesandbox.io/s/github/IgniteUI/igniteui-blazor-examples/tree/master/samples/grids/grid/editing-excel-style-custom?fontsize=14&hidenavigation=1&theme=dark&view=preview&file=/src/App.razor" rel="noopener noreferrer">
-            <img height="40px" style="border-radius: 0rem; max-width: 100%;" alt="Run Sample" src="https://github.com/IgniteUI/igniteui-blazor-examples/raw/vnext/templates/sample/images/button-sandbox.png"/>
-        </a> -->
-    </body>
-</html>
-
-## Branches
-
-> **_NOTE:_** You should use [master](https://github.com/IgniteUI/igniteui-blazor-examples/tree/master) branch of this repository if you want to run samples on your computer. Use the [vnext](https://github.com/IgniteUI/igniteui-blazor-examples/tree/vnext) branch only when you want to contribute new samples to this repository.
-
-## Setup
-
-- instal **.NET SDK** from this [website](https://dotnet.microsoft.com/learn/aspnet/blazor-tutorial/install)
-
-## Running App in Visual Studio 2019
-
-NOTE: VS 2019 has better code highlighting and error detection than VS Code does.
-
-- open **Visual Studio 2019** as an administrator
-
-- open the **BlazorClientApp.sln** solution
-
-- right click solution and select **Restore NuGet Packages** menu item
-
-- click **Debug** menu and select **Start Debugging** or press **F5** key
-
-
-## Running App in VS Code
-
-- open **VS Code** as an administrator
-
-- open this folder in **VS Code**
-
-- open a terminal window
-
-- to restore assemblies, run this command:
-```dotnet restore```
-
-- to run samples, run this command:
-```dotnet watch run```
-
-- wait for for message:
-**Now listening on: http://localhost:4200**
-
-- open **http://localhost:4200** in your browser
-
-
-## Resources
-
-- [Razor Component Models](https://www.codemag.com/article/1911052)
-- [Razor Syntax](https://docs.microsoft.com/en-us/aspnet/core/blazor/components/?view=aspnetcore-3.1#razor-syntax)
-- [Getting reference to components](https://docs.microsoft.com/en-us/aspnet/core/blazor/components/?view=aspnetcore-3.1#capture-references-to-components)
