--- samples/grids/grid/editing-excel-style-custom/_Imports.razor
@@ -1,9 +1,0 @@
-// these namespaces are global to the app
-@using System.Net.Http
-@using System.Net.Http.Json
-@using Microsoft.AspNetCore.Components.Forms
-@using Microsoft.AspNetCore.Components.Routing
-@using Microsoft.AspNetCore.Components.Web
-@using Microsoft.AspNetCore.Components.WebAssembly.Http
-@using Microsoft.JSInterop
-@using Infragistics.Samples
