--- samples/grids/grid/editing-excel-style-custom/wwwroot/index.css
@@ -1,4 +1,0 @@
-/*
-CSS styles are loaded from the shared CSS file located at:
-https://static.infragistics.com/xplatform/css/samples/
-*/
