+++ samples/grids/grid/editing-excel-style/wwwroot/events.js
@@ -1,35 +1,79 @@
-
-
 igRegisterScript("WebGridEditingExcelStyle", (ev) => {
-    var key = ev.detail.event.keyCode;
-    var grid = ev.detail.target.grid;
-    var activeElem = grid.navigation.activeNode;
+    // Assuming `ev.detail.target` gives you the grid element or another relevant element
+    var gridElement = ev.detail.target; // This is just an example; adjust based on actual structure
 
-    if ((key >= 48 && key <= 57) || (key >= 65 && key <= 90) || (key >= 97 && key <= 122)) {
-        var columnName = grid.getColumnByVisibleIndex(activeElem.column).field;
-        var cell = grid.getCellByColumn(activeElem.row, columnName);
+    // Attach the onkeydown event listener to the gridElement or a specific child element
+    gridElement.addEventListener('keydown', (keydownEvent) => {
 
-        if (cell && !grid.crudService.cellInEditMode) {
-            grid.crudService.enterEditMode(cell);
-            cell.editValue = key;
+    }, false);
+
+});
+
+var shouldAppendValue = false;
+
+window.attachKeyDownEvent = () => {
+    var grid1 = document.querySelectorAll("igc-grid")[0];
+
+    grid1.addEventListener('activeNodeChange', (event) => {
+        grid1.endEdit();
+        (grid1.getElementsByClassName("igx-grid__tbody-content")[0]).focus();
+    });
+
+    grid1.addEventListener('keydown', (e) => {
+        var code = e.code;
+        var grid1 = window.grid1;
+        var activeElem = grid1.selectedCells[0];
+
+        if ((e.code >= 'Digit0' && e.code <= 'Digit9') ||
+            (e.code >= 'KeyA' && e.code <= 'KeyZ') ||
+            (e.code >= 'Numpad0' && e.code <= 'Numpad9') &&
+             e.code !== 'Enter' && e.code !== 'NumpadEnter') {
+
+            if (activeElem && activeElem.editMode === false) {
+                activeElem.editMode = true;
+                activeElem.editValue = e.key;
+                shouldAppendValue = true;
+                grid1.markForCheck();
+            } else
+
+                if (activeElem && activeElem.editMode && this.shouldAppendValue) {
+                    e.preventDefault();
+                    activeElem.editValue = activeElem.editValue + e.key;
+                    shouldAppendValue = false;
+                }
         }
-    }
 
-    if (key == 13) {
-        var thisRow = activeElem.row;
-        var column = activeElem.column;
-        var rowInfo = grid.dataView;
+        if (code === 'Backspace') {
+            if (activeElem == null || activeElem.editMode == false) {
+                return;
+            }
+            const rowIndex = activeElem.row.index;
+            const columnKey = activeElem.column.field;
 
-        var nextRow = this.getNextEditableRowIndex(thisRow, rowInfo, ev.detail.event.shiftKey);
+            grid1.data[rowIndex][columnKey] = '';
+            grid1.markForCheck();
 
-        grid.navigateTo(nextRow, column, (obj) => {
-            obj.target.activate();
-            grid.clearCellSelection();
-        });
-    }
-}, false);
+        }
 
-function getNextEditableRowIndex(currentRowIndex, dataView, previous) {
+        if (code === 'Enter' || code === 'NumpadEnter') {
+
+            if (activeElem == null) {
+                return;
+            }
+
+            const thisRow = activeElem.row.index;
+            const dataView = this.grid1.dataView;
+            const nextRowIndex = this.getNextEditableRowIndex(thisRow, dataView, e.shiftKey);
+
+            grid1.navigateTo(nextRowIndex, activeElem.column.visibleIndex, (obj) => {
+                grid1.clearCellSelection();
+                obj.target.activate();
+            });
+        }
+    });
+}
+
+window.getNextEditableRowIndex = (currentRowIndex, dataView, previous) => {
     if (currentRowIndex < 0 || (currentRowIndex === 0 && previous) || (currentRowIndex >= dataView.length - 1 && !previous)) {
         return currentRowIndex;
     }
