+++ samples/grids/grid/editing-excel-style/wwwroot/index.html
@@ -26,6 +26,8 @@
     <!-- importing blazor bundle for IG controls: -->
     <script src="_content/IgniteUI.Blazor/app.bundle.js"></script>
     <script src="_framework/blazor.webassembly.js"></script>
+    <Routes @rendermode=RenderMode.InteractiveServer />
+    <script src="_framework/blazor.web.js"></script>
     <!-- importing java scripts used in razor page -->
     <script src="events.js"></script>
 </body>
